Update hash method and writing conditions

+++ src/context/block_unit.rs
@@ -1833,7 +1833,7 @@ impl ContextWriter<'_> {
       //use log::info;
       //info!("Used a hash");
       // PERF: We can encode this in a more efficient manner
-      for byte in (hash as u16).to_be_bytes() {
+      for byte in (hash as u32).to_be_bytes() {
         w.bit(((byte >> 7) & 0b1).into());
         w.bit(((byte >> 6) & 0b1).into());
         w.bit(((byte >> 5) & 0b1).into());
